+++ libs/api-service/README.md
@@ -1,0 +1,7 @@
+# api-service
+
+This library was generated with [Nx](https://nx.dev).
+
+## Building
+
+Run `nx build api-service` to build the library.
