+++ libs/api-service/package.json
@@ -1,0 +1,10 @@
+{
+  "name": "@apiService",
+  "version": "0.0.1",
+  "dependencies": {
+    "tslib": "^2.3.0"
+  },
+  "type": "commonjs",
+  "main": "./src/index.js",
+  "typings": "./src/index.d.ts"
+}
