+++ libs/api-service/project.json
@@ -1,0 +1,26 @@
+{
+  "name": "api-service",
+  "$schema": "../../node_modules/nx/schemas/project-schema.json",
+  "sourceRoot": "libs/api-service/src",
+  "projectType": "library",
+  "targets": {
+    "build": {
+      "executor": "@nx/js:tsc",
+      "outputs": ["{options.outputPath}"],
+      "options": {
+        "outputPath": "dist/libs/api-service",
+        "main": "libs/api-service/src/index.ts",
+        "tsConfig": "libs/api-service/tsconfig.lib.json",
+        "assets": ["libs/api-service/*.md"]
+      }
+    },
+    "lint": {
+      "executor": "@nx/linter:eslint",
+      "outputs": ["{options.outputFile}"],
+      "options": {
+        "lintFilePatterns": ["libs/api-service/**/*.ts", "libs/api-service/package.json"]
+      }
+    }
+  },
+  "tags": []
+}
