+++ libs/interface/README.md
@@ -1,0 +1,7 @@
+# interface
+
+This library was generated with [Nx](https://nx.dev).
+
+## Building
+
+Run `nx build interface` to build the library.
