+++ libs/interface/project.json
@@ -1,0 +1,26 @@
+{
+  "name": "interface",
+  "$schema": "../../node_modules/nx/schemas/project-schema.json",
+  "sourceRoot": "libs/interface/src",
+  "projectType": "library",
+  "targets": {
+    "build": {
+      "executor": "@nx/js:tsc",
+      "outputs": ["{options.outputPath}"],
+      "options": {
+        "outputPath": "dist/libs/interface",
+        "main": "libs/interface/src/index.ts",
+        "tsConfig": "libs/interface/tsconfig.lib.json",
+        "assets": ["libs/interface/*.md"]
+      }
+    },
+    "lint": {
+      "executor": "@nx/linter:eslint",
+      "outputs": ["{options.outputFile}"],
+      "options": {
+        "lintFilePatterns": ["libs/interface/**/*.ts", "libs/interface/package.json"]
+      }
+    }
+  },
+  "tags": []
+}
