+++ libs/interface/src/Request.ts
@@ -1,0 +1,9 @@
+import { Network } from '@config';
+
+export type ClaimParams = {
+  address: string;
+  message: string;
+  signature: string;
+  captcha: string;
+  network?: Network;
+};
