+++ tsconfig.base.json
@@ -16,10 +16,12 @@
     "baseUrl": ".",
     "paths": {
       "@/*": ["./apps/web/*"],
+      "@apiService": ["libs/api-service/src/index.ts"],
       "@blockchainService": ["libs/blockchain-service/src/index.ts"],
       "@config": ["libs/config/src/index.ts"],
       "@errors": ["libs/errors/src/index.ts"],
       "@faucetService": ["libs/faucet-service/src/index.ts"],
+      "@interface": ["libs/interface/src/index.ts"],
       "@nonceService": ["libs/nonce-service/src/index.ts"],
       "@securityService": ["libs/security-service/src/index.ts"],
       "@trackingService": ["libs/tracking-service/src/index.ts"],
